Add priority to instances in object and db
Add string column "priority" to the instances table in DB and
corresponding field in the Instance versioned object. Add db and
versioned object operations accordingly.

The "priority" is used to identify the high or low priority of the
instances. If priority is not set, the current instance is a common
instance with no priority.